+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Display "Inactive" by inactive locations on instance view. Fixes UIIN-1970.
 * Create a new instance search option. Fixes UIIN-2264.
 * Add create/update widget to HRID Settings page. UIIN-2139.
+* Do not fetch bound with data if the boundWithParts data is not present. Fixes UIIN-2272.
 
 ## [9.2.0](https://github.com/folio-org/ui-inventory/tree/v9.2.0) (2022-10-27)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.1.0...v9.2.0)